Сдаём задачу спринта №22 на ревью

+++ app/src/main/java/com/practicum/playlistmaker/media_library/domain/api/FavoriteTrackInteractor.kt
@@ -11,5 +11,5 @@ interface FavoriteTrackInteractor {
 
     suspend fun getFavoriteTracks(): Flow<List<Track>>
 
-    suspend fun getFavoriteIdList(): List<Int>
+    suspend fun getFavoriteIdList(): List<Long>
 }
